+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ### Added
 
+- Added `blobSoftDeleteEnabled` and `blobSoftDeleteRetentionDays` to
+  `azure_storage_account`
 - Added `networkRuleSetDefaultAction` and `networkRuleSetBypass` properties to
   `azure_storage_account`.